game_logic/requirements.txt
pygame-ce
pyserial
Nuitka